# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhereWeAre API",
    "version": "1.0.0",
    "description": "Comprehensive API for accessing cultural landmarks, user management, memories, and favorites in Chemnitz"
  },
  "servers": [
    {
      "url": "https://whereweare.vercel.app",
      "description": "Application server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
      "CulturalSite": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier"
          },
          "osmId": {
            "type": "string",
            "description": "OpenStreetMap ID"
          },
          "name": {
            "type": "string",
            "description": "Name of the cultural site"
          },
          "category": {
            "type": "string",
            "description": "Category of the cultural site"
          },
          "description": {
            "type": "string",
            "description": "Description of the site"
          },
          "address": {
            "type": "string",
            "description": "Street address"
          },
          "coordinates": {
            "type": "object",
            "properties": {
              "lat": {
                "type": "number",
                "description": "Latitude"
              },
              "lng": {
                "type": "number",
                "description": "Longitude"
              }
            }
          },
          "tags": {
            "type": "object",
            "description": "OpenStreetMap tags"
          },
          "website": {
            "type": "string",
            "description": "Official website URL"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number"
          },
          "openingHours": {
            "type": "string",
            "description": "Opening hours information"
          },
          "accessibility": {
            "type": "object",
            "properties": {
              "wheelchair": {
                "type": "string",
                "enum": [
                  "yes",
                  "no",
                  "limited"
                ]
              },
              "parking": {
                "type": "boolean"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CulturalSite"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/cultural-sites/nearby": {
      "get": {
        "summary": "Get cultural sites near a location",
        "description": "Find cultural sites within a specified radius of given coordinates. All query parameters are received as strings and parsed to numbers.",
        "tags": [
          "Cultural Sites"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "type": "string",
              "format": "float"
            },
            "description": "Latitude of the center point"
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "string",
              "format": "float"
            },
            "description": "Longitude of the center point"
          },
          {
            "in": "query",
            "name": "radius",
            "schema": {
              "type": "string",
              "format": "float",
              "default": "5"
            },
            "description": "Search radius in kilometers"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string",
              "format": "integer",
              "default": "20",
              "maximum": 100
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "List of nearby cultural sites with distances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/CulturalSite"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "distance": {
                            "type": "number",
                            "description": "Distance in kilometers"
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing or invalid coordinates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/cultural-sites": {
      "get": {
        "summary": "Get cultural sites with filtering and pagination",
        "description": "Get cultural sites with optional filtering and pagination. All query parameters are received as strings and parsed internally.",
        "tags": [
          "Cultural Sites"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string",
              "format": "integer",
              "default": "1"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string",
              "format": "integer",
              "default": "20",
              "maximum": 1000
            },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category (use /api/cultural-sites/categories to get available categories)"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search term for site names and descriptions"
          },
          {
            "in": "query",
            "name": "north",
            "schema": {
              "type": "string",
              "format": "float"
            },
            "description": "Northern boundary for geographic filtering"
          },
          {
            "in": "query",
            "name": "south",
            "schema": {
              "type": "string",
              "format": "float"
            },
            "description": "Southern boundary for geographic filtering"
          },
          {
            "in": "query",
            "name": "east",
            "schema": {
              "type": "string",
              "format": "float"
            },
            "description": "Eastern boundary for geographic filtering"
          },
          {
            "in": "query",
            "name": "west",
            "schema": {
              "type": "string",
              "format": "float"
            },
            "description": "Western boundary for geographic filtering"
          },
          {
            "in": "query",
            "name": "wheelchair",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Filter for wheelchair accessible sites (converted to accessibility.wheelchair = \"yes\")"
          },
          {
            "in": "query",
            "name": "parking",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Filter for sites with parking"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of cultural sites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/memories/{id}/image/{imageId}": {
      "get": {
        "summary": "Get a memory image",
        "description": "Retrieve a specific image from a memory by ID. Returns raw image data with appropriate Content-Type header, caching headers, and CORS support.",
        "tags": [
          "Memories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Memory ID"
          },
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image ID within the memory"
          }
        ],
        "responses": {
          "200": {
            "description": "Image file",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Type": {
                "description": "Image MIME type",
                "schema": {
                  "type": "string",
                  "example": "image/jpeg"
                }
              },
              "Content-Length": {
                "description": "Image size in bytes",
                "schema": {
                  "type": "integer"
                }
              },
              "Cache-Control": {
                "description": "Caching directives",
                "schema": {
                  "type": "string",
                  "example": "public, max-age=31536000, immutable"
                }
              }
            }
          },
          "404": {
            "description": "Memory or image not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}